Refactor integ test
